[#noissue] Extract OTel ExceptionInfo resolution into OtlpExceptionInfoResolver
